requirements.txt
yfinance>=0.2.0
pandas>=1.5.0
numpy>=1.20.0
google-generativeai>=0.3.0
scikit-learn>=1.0.0
streamlit>=1.28.0
bcrypt>=4.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0

